Add a basic implementation of relational algebra
This is a WIP implementation of relational algebra that runs in Hydroflow.

Right now it only uses scheduled components and is only at runtime, no codegen,
just for simplicity. It's also not how I would design a non-proof-of-concept IR
but I think for our purposes for now it will be fine.