Add a simple codegen path for relational algebra (#24)
This currenly only handles some very simple patterns, and doesn't invoke
scheduled components at all, but this is the general pattern I was thinking for
now. Comments on the overall design welcome, you can see examples of the
generated code in testdata/compile/compile.

What's more, I think this PR will probably fail on CI since it shells out to
rustfmt to format the generated code. Suggestions welcome for how to fix, I
might just make it skip running the test (probably with a warning) if rustfmt
does not just work.

Next step is to introduce a let binding layer to the language, to force the
codegen to use scheduled components/invoke Hydroflow.