Switch from a bespoke format -> bincode (#40)
This commit tidies up the TCP stuff more. There's still more to be done in
terms of ergonomics but I want to wait for the API change to work on that. I
think the next steps are that, and also writing a benchmark against a
hand-coded server, since I think there's a lot of low-hanging fruit in terms of
stuff like buffer reuse.