Fix some Compare implementations
The Compare implementations for DomPair and MapUnion were incorrect.  I'd like
to add some tests around this but I think maybe there's something more holistic
to be done here, given all the algebraic properties we have of lattices this
seems like it should be very amenable to property-based-testing.

The MapUnion implementation in particular feels very...icky. We probably want
something a bit more specialized for various tags that permit a better
implementation, like exploiting order or fast random lookups. But maybe there's
a nicer way to implement it here? Not sure, couldn't think of anything.