Add another implementation of KVS (#122)
This one doesn't use the surface API directly, and just uses the scheduled
layer. It gets about the same performance, and flame graphs for both
implementations show a lot of time spent in tokio scheduling.  I'm going to
bring back the old non-hydroflow version and see what kind of perf that gets,
for comparison.