Make KVS benchmark more Anna-like (#146)
I decided it made the most sense to remove the other implementations here,
since there were a lot of structural changes going on, they made it annoying to
update this. They live on in the repo history if they ever need to be revived.

This commit moves the workload generation into the workers themselves, since
otherwise I was unable to fully saturate the workers.