Add a driver for the chat example
Clients can take messages in on stdin from any process, but in order to drive them autonomously we
need a reliable and continuous source of input that's less spammy than commands like 'yes'. This
diff adds a chat driver that spits out a random phrase every few seconds to serve as this driver.