Refining the examples container's definition
In order for the chat driver to run from within the examples container, I needed to either install
Python in the container or rewrite the driver in Rust. I opted to install Python, mainly in the
interest of getting something working end-to-end.

The rust:slim base image doesn't ship a dictionary, so I've checked in a local copy of
`/usr/share/dict/web2a` that the chat driver can use.

I've also moved the chat driver out into `example_utils/`, rather than keeping it in `examples/`,
since it makes the copying logic between container build stages a little easier to express.