Allow the chat example to bind and connect to DNS names
In many circumstances when testing chat in a distributed setting, we won't know the right IP address
to either bind or connect to in advance. This diff allows the chat client and server to bind to
DNS names, which (coupled with the distributed runtime's service discovery mechanism) should make it
easier for clients and servers to find each other.