Refactor `OpConstraints` fns, use lookup helper (#399)
Replaces `dyn Fn()` (dynamic dispatch trait) with `fn()` (fn pointer).
The main difference is that `dyn Fn()` has a notion of `&self` while
`fn()` is a pure function pointer. We don't use `&self` in any of the op
fns so `fn` makes more sense (and is `Send + Sync`).

Also use lookup helpers to avoid `.find()`ing in the `OPERATORS` slice.