feat(hydroflow_plus): provide simpler API for launching and minimize dependencies
The new `hydroflow::cli::util::launch` API takes a function from the `HydroCLI` config to a `Hydroflow`, which allows us to wait until the `Hydroflow` struct is instantiated (and all network connections are established) before returning the "ack start" message. This is also why this PR is stacked, because it triggers the race condition much more consistently.

This also disables depending on default features of Hydroflow from Hydroflow+, which should speed up builds because we no longer load the Hydroflow / Datalog macros.