feat(anna): Added a two-level data model for the layer zero key-value store.
The data model supported by the KV store is analogous to the following JSON:

```json
{
   "sys": [
      "members": [
              "key": "value"
      ]
   ],
   "usr": [
       "table_a": [
           ...
       ],
       "table_b": [
           ...
       ],

   ]
}
```

A single `TableMap` can be used to model `sys` or `usr`. We're only planning for these two namespaces for now, so there is no higher-level lattices to allow for more namespaces.

Initially, I started with support for deleting an entire table efficiently: but I have decided to drop that until the need arises. If it isn't needed in practice, the lack of a sharp operation that can delete an entire table leads to a safer system. It also simplified the code a fair bit.

The tests may feel like they're testing lattice merging (and therefore redundant), but I wrote these to convince myself that type signatures of the data model are working. I may remove them in the future (or not!).