feat(anna): Escaping of keys (#1265)
It was not the most critical thing to get in, but it bothered me a bit.
This code change allows '/' to be used in table and row names. This
doesn't affect memory layout in any way.

The most critical part of this change is the additional tests to check
for empty table names and row keys.