feat(anna): Deterministic testing for L0 server (and also, return multiple values for concurrent writes) (#1275)
These changes separate the server logic from the networking logic,
allowing the server logic to be tested with high fidelity in a unit
test. The changes allow for an abstract replacement of `SocketAddr`,
which can be used for deterministic testing of more complex scenarios in
unit tests.

The first feature to use these testing capabilities is concurrent writes
to the same key at the same tick, which should result in multiple values
being returned from a `Get`.