feat(hydroflow_plus): add ability to have staged flows inside unit tests
Whenever a Hydroflow+ program is compiled, it depends on a generated `__staged` module, which contains the entire contents of the crate but with every type / function made `pub` and exported, so that the compiled UDFs can resolve local references appropriately.

Previously, we would not do this for `#[cfg(test)]` modules, since they may use `dev-dependencies` and therefore the generated module may fail to compile when not in test mode. To solve this, when running a unit test (marked with `hydroflow_plus::deploy::init_test()`) that uses trybuild, we emit a version of the `__staged` module with `#[cfg(test)]` modules included _into the generated trybuild sources_ because we can guarantee via trybuild that the appropriate `dev-dependencies` are available.

This by itself allows crates depending on `hydroflow_plus` to have local unit tests with Hydroflow+ logic inside them. But we also want to use this support for unit tests inside `hydroflow_plus` itself. To enable that, we eliminate the `hydroflow_plus_deploy` crate and move its contents directly to `hydroflow_plus` itself so that we can access the trybuild machinery without incurring a circular dependency.

Also fixes #1408