feat(anna): Get minikube working (#1472)
When I started, I was using the published Hydroflow. I have since
switched to using local hydroflow, which means the Dockerfile needs to
be updated for the build to work correctly. This posed a more
significant challenge than expected, primarily because of disk space
issues, which have been sorted out.