docs: README.md hydroflow+ => hydro, hydroflow => DFIR, rewording (#1621)
Here's a version of the README that has reasonable text with the name
changes. The links will need to be updated to point to changes names for
the relevant directories/files in the docs (which should be changed
appropriately).

Feel free to copy this file and reject this PR, or merge this PR and
rebase, as you see fit.

Mingwei: also adds basic redirects for the old paths

---------

Co-authored-by: Mingwei Samuel <mingwei.samuel@gmail.com>